created product and rating component, and rating stars logic
