+++ frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\reportWebVitals.js":"1","C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\App.js":"2","C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\data.js":"3","C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\index.js":"4"},{"size":362,"mtime":1606652038246,"results":"5","hashOfConfig":"6"},{"size":1870,"mtime":1610367016771,"results":"7","hashOfConfig":"6"},{"size":1341,"mtime":1610366012037,"results":"8","hashOfConfig":"6"},{"size":500,"mtime":1606652038246,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1kigm9i",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\reportWebVitals.js",[],"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\App.js",[],"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\data.js",[],"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\index.js",[]]
+[{"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\reportWebVitals.js":"1","C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\data.js":"2","C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\App.jsx":"3","C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\components\\Product.jsx":"4","C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\components\\Rating.jsx":"5"},{"size":362,"mtime":1606652038246,"results":"6","hashOfConfig":"7"},{"size":1336,"mtime":1610370043938,"results":"8","hashOfConfig":"7"},{"size":831,"mtime":1610369678474,"results":"9","hashOfConfig":"7"},{"size":708,"mtime":1610369666444,"results":"10","hashOfConfig":"7"},{"size":1412,"mtime":1610369942605,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1kigm9i",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\reportWebVitals.js",[],"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\data.js",[],"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\App.jsx",[],"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\components\\Product.jsx",[],"C:\\Users\\victo\\Documents\\GitHub\\E-commerce-REACT-REDUX\\frontend\\src\\components\\Rating.jsx",[]]
