+++ frontend/src/data.js
@@ -7,7 +7,7 @@ const data ={
             image: '/images/p1.jpg',
             price: 350,
             brand: 'Conor',
-            rating: 4.5,
+            rating: 3.5,
             numReviews: 10,
             description: 'high quality product',
           },
@@ -19,7 +19,7 @@ const data ={
             price: 2999,
             brand: 'Felt',
             rating: 4.5,
-            numReviews: 10,
+            numReviews: 17,
             description: 'high quality product',
           },
         {
@@ -29,8 +29,8 @@ const data ={
             image: '/images/p3.jpg',
             price: 1475,
             brand: 'merida',
-            rating: 4.5,
-            numReviews: 10,
+            rating: 4,
+            numReviews: 24,
             description: 'high quality product',
           },
         {
@@ -40,8 +40,8 @@ const data ={
             image: '/images/p4.jpg',
             price: 2875,
             brand: 'Scott',
-            rating: 4.5,
-            numReviews: 10,
+            rating: 5,
+            numReviews: 8,
             description: 'high quality product',
           },
     ]
