+++ frontend/src/index.css
@@ -112,3 +112,6 @@ color: #f0c040;
 margin: 0.1rem;
 }
 
+.rating span:last-child {
+  color: #404040;
+}
